+++ float_math.h
@@ -1,7 +1,7 @@
 #ifndef FLOAT_MATH_H
 #define FLOAT_MATH_H
 
-int isEqual(float a, float b);
-int isZero(float a);
+int isEqual(double a, double b);
+int isZero(double a);
 
 #endif
